refactored evidence levels for teachers and students
