+++ apps/src/templates/rubrics/EvidenceLevels.jsx
@@ -1,16 +1,8 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import classNames from 'classnames';
-import i18n from '@cdo/locale';
-import style from './rubrics.module.scss';
 import {evidenceLevelShape, submittedEvaluationShape} from './rubricShapes';
-import RadioButton from '@cdo/apps/componentLibrary/radioButton/RadioButton';
-import {
-  BodyThreeText,
-  Heading6,
-  StrongText,
-} from '@cdo/apps/componentLibrary/typography';
-import {UNDERSTANDING_LEVEL_STRINGS} from './rubricHelpers';
+import EvidenceLevelsForStudents from './EvidenceLevelsForStudents';
+import EvidenceLevelsForTeachers from './EvidenceLevelsForTeachers';
 
 export default function EvidenceLevels({
   evidenceLevels,
@@ -25,61 +17,20 @@ export default function EvidenceLevels({
     return newArray.sort((a, b) => b.understanding - a.understanding);
   };
   if (canProvideFeedback) {
-    const radioGroupName = `evidence-levels-${learningGoalKey}`;
     return (
-      <div className={style.evidenceLevelSet}>
-        <Heading6>{i18n.assignARubricScore()}</Heading6>
-        {sortedEvidenceLevels().map(evidenceLevel => (
-          <div
-            key={evidenceLevel.id}
-            className={classNames(
-              style.evidenceLevelOption,
-              style.evidenceLevelLabel
-            )}
-          >
-            {' '}
-            <RadioButton
-              label={UNDERSTANDING_LEVEL_STRINGS[evidenceLevel.understanding]}
-              name={radioGroupName}
-              value={evidenceLevel.id}
-              size="s"
-              onChange={() => {
-                radioButtonCallback(evidenceLevel.understanding);
-              }}
-              checked={understanding === evidenceLevel.understanding}
-            />
-            <BodyThreeText
-              className={classNames(style.evidenceLevelDescriptionIndented)}
-            >
-              {evidenceLevel.teacherDescription}
-            </BodyThreeText>
-          </div>
-        ))}
-      </div>
+      <EvidenceLevelsForTeachers
+        learningGoalKey={learningGoalKey}
+        evidenceLevels={sortedEvidenceLevels()}
+        understanding={understanding}
+        radioButtonCallback={radioButtonCallback}
+      />
     );
   } else {
     return (
-      <div className={style.evidenceLevelSet}>
-        <Heading6>{i18n.rubricScores()}</Heading6>
-        {sortedEvidenceLevels().map(evidenceLevel => (
-          <div
-            key={evidenceLevel.id}
-            className={classNames(style.evidenceLevelOption, {
-              [style.submittedEvaluationEvidenceLevel]:
-                submittedEvaluation?.understanding ===
-                evidenceLevel.understanding,
-            })}
-          >
-            {/*TODO: [DES-321] Label-two styles here*/}
-            <BodyThreeText>
-              <StrongText>
-                {UNDERSTANDING_LEVEL_STRINGS[evidenceLevel.understanding]}
-              </StrongText>
-            </BodyThreeText>
-            <BodyThreeText>{evidenceLevel.teacherDescription}</BodyThreeText>
-          </div>
-        ))}
-      </div>
+      <EvidenceLevelsForStudents
+        evidenceLevels={sortedEvidenceLevels()}
+        submittedEvaluation={submittedEvaluation}
+      />
     );
   }
 }
