+++ apps/test/unit/templates/rubrics/EvidenceLevelsForTeachersTest.jsx
@@ -1,0 +1,50 @@
+import React from 'react';
+import {expect} from '../../../util/reconfiguredChai';
+import {shallow, mount} from 'enzyme';
+import sinon from 'sinon';
+import EvidenceLevelsForTeachers from '@cdo/apps/templates/rubrics/EvidenceLevelsForTeachers';
+import {UNDERSTANDING_LEVEL_STRINGS} from '@cdo/apps/templates/rubrics/rubricHelpers';
+
+const DEFAULT_PROPS = {
+  evidenceLevels: [
+    {id: 1, understanding: 0, teacherDescription: 'test1'},
+    {id: 2, understanding: 1, teacherDescription: 'test2'},
+  ],
+  learningGoalKey: 'key-1',
+};
+
+describe('EvidenceLevels', () => {
+  it('renders evidence levels', () => {
+    const wrapper = shallow(<EvidenceLevelsForTeachers {...DEFAULT_PROPS} />);
+    expect(wrapper.find('Heading6').length).to.equal(1);
+    expect(wrapper.find('Heading6').props().children).to.equal(
+      'Assign a Rubric Score'
+    );
+    expect(wrapper.find('Memo(RadioButton)').length).to.equal(
+      DEFAULT_PROPS.evidenceLevels.length
+    );
+    expect(wrapper.find('BodyThreeText').length).to.equal(
+      DEFAULT_PROPS.evidenceLevels.length
+    );
+    const firstEvidenceLevel = DEFAULT_PROPS.evidenceLevels[0];
+    expect(wrapper.find('BodyThreeText').at(0).props().children).to.equal(
+      firstEvidenceLevel.teacherDescription
+    );
+    expect(wrapper.find('Memo(RadioButton)').at(0).prop('label')).to.equal(
+      UNDERSTANDING_LEVEL_STRINGS[firstEvidenceLevel.understanding]
+    );
+  });
+
+  it('calls radioButtonCallback when understanding is selected', () => {
+    const callback = sinon.stub();
+    const wrapper = mount(
+      <EvidenceLevelsForTeachers
+        {...DEFAULT_PROPS}
+        radioButtonCallback={callback}
+      />
+    );
+    wrapper.find('input').first().simulate('change');
+    sinon.assert.calledOnce(callback);
+    wrapper.unmount();
+  });
+});
