+++ apps/test/unit/templates/rubrics/EvidenceLevelsTest.jsx
@@ -1,9 +1,7 @@
 import React from 'react';
 import {expect} from '../../../util/reconfiguredChai';
-import {shallow, mount} from 'enzyme';
-import sinon from 'sinon';
+import {shallow} from 'enzyme';
 import EvidenceLevels from '@cdo/apps/templates/rubrics/EvidenceLevels';
-import {UNDERSTANDING_LEVEL_STRINGS} from '@cdo/apps/templates/rubrics/rubricHelpers';
 
 const DEFAULT_PROPS = {
   evidenceLevels: [
@@ -14,62 +12,17 @@ const DEFAULT_PROPS = {
 };
 
 describe('EvidenceLevels', () => {
-  it('renders evidence levels when feedback available', () => {
+  it('renders teachers view of evidence levels when the user can provide feedback', () => {
     const wrapper = shallow(
       <EvidenceLevels {...DEFAULT_PROPS} canProvideFeedback />
     );
-    expect(wrapper.find('Heading6').length).to.equal(1);
-    expect(wrapper.find('Heading6').props().children).to.equal(
-      'Assign a Rubric Score'
-    );
-    expect(wrapper.find('Memo(RadioButton)').length).to.equal(
-      DEFAULT_PROPS.evidenceLevels.length
-    );
-    expect(wrapper.find('BodyThreeText').length).to.equal(
-      DEFAULT_PROPS.evidenceLevels.length
-    );
-    const lastEvidenceLevel =
-      DEFAULT_PROPS.evidenceLevels[DEFAULT_PROPS.evidenceLevels.length - 1];
-    expect(wrapper.find('BodyThreeText').at(0).props().children).to.equal(
-      lastEvidenceLevel.teacherDescription
-    );
-    expect(wrapper.find('Memo(RadioButton)').at(0).prop('label')).to.equal(
-      UNDERSTANDING_LEVEL_STRINGS[lastEvidenceLevel.understanding]
-    );
+    expect(wrapper.find('EvidenceLevelsForTeachers').length).to.equal(1);
   });
 
-  it('renders evidence levels when feedback not available', () => {
+  it('renders student view of evidence levels when feedback not available', () => {
     const wrapper = shallow(
       <EvidenceLevels {...DEFAULT_PROPS} canProvideFeedback={false} />
     );
-    expect(wrapper.find('Heading6').length).to.equal(1);
-    expect(wrapper.find('Heading6').props().children).to.equal('Rubric Scores');
-    expect(wrapper.find('Memo(RadioButton)').length).to.equal(0);
-    // Two BodyThreeText per evidence level
-    expect(wrapper.find('BodyThreeText').length).to.equal(
-      DEFAULT_PROPS.evidenceLevels.length * 2
-    );
-    const lastEvidenceLevel =
-      DEFAULT_PROPS.evidenceLevels[DEFAULT_PROPS.evidenceLevels.length - 1];
-    expect(wrapper.find('StrongText').at(0).props().children).to.equal(
-      UNDERSTANDING_LEVEL_STRINGS[lastEvidenceLevel.understanding]
-    );
-    expect(wrapper.find('BodyThreeText').at(1).props().children).to.equal(
-      lastEvidenceLevel.teacherDescription
-    );
-  });
-
-  it('calls radioButtonCallback when understanding is selected', () => {
-    const callback = sinon.stub();
-    const wrapper = mount(
-      <EvidenceLevels
-        {...DEFAULT_PROPS}
-        canProvideFeedback
-        radioButtonCallback={callback}
-      />
-    );
-    wrapper.find('input').first().simulate('change');
-    sinon.assert.calledOnce(callback);
-    wrapper.unmount();
+    expect(wrapper.find('EvidenceLevelsForStudents').length).to.equal(1);
   });
 });
